+++ CHANGELOG.md
@@ -20,12 +20,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SnapshotTesting dependency.
 - `onFirstAppear` view modifier.
 - Hero Detail previews.
+- Kingfisher dependency to Swift package.
+- Hero Detail feature presentation.
 
 ### Changed
 
 - Bump Deployment Target to iOS 15.
 - Bump Swift Language Version to 6.0.
 
+### Removed
+
+- Kingfisher dependency from Podfile.
+
 ### Fixed
 
 - Kingfisher build [issue](https://github.com/onevcat/Kingfisher/issues/2052).
+- Concurrency warnings.